==> scripts/05_chatgpt_tagging/12_output.R <==

# Save final data from chatgpt_tagging pipeline

# Store data locally ------------------------------------------------------

save(all_businesses_eval, file = here("scripts", "05_chatgpt_tagging", "chatgpt_tagging_output", "all_businesses_eval.RData"))
